requirements.txt
flask
flask-cors
Pillow
numpy<2
torch==2.2.1+cpu
torchvision==0.17.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu